web_ui/pyproject.toml
[project]
name = "augment-lite-web-ui"
version = "0.7.0"
description = "Web UI for augment-lite MCP server with real-time logs and configuration management"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.121.1",
    "uvicorn[standard]==0.34.0",
    "websockets==14.1",
    "jinja2==3.1.5",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "aiofiles==24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "httpx==0.28.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "httpx>=0.28.1",
]
